Lift ownership of filtered site data up into <SitesDashboard>
